backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
peewee>=3.17.0
requests>=2.31.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
PyJWT>=2.8.0
